use smaller type for UAudioManager::_granularity
There is no point in using an unsigned long if timestamps returned
by fmod are unsigned int and the result of our subtraction is always
positive.

+++ src/uAudioManager.h
@@ -80,7 +80,7 @@ private:
     UAudioManager();
     WidgetSongData * _widgetSongData;
     unsigned int _lastPosition;
-    unsigned long _granularity;
+    unsigned int _granularity;
     qint64 _delta;
     double _speedFactor;
     float _frequency;
